rag-langFlow-mcp/requirements.txt
# LangChain and RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ibm>=0.1.0

# Document processing
pypdf>=3.17.0

# Vector database
pgvector>=0.2.4
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# LLM providers
openai>=1.12.0
tiktoken>=0.5.0

# MCP Server
fastmcp>=0.2.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
